gun: merge mesh utilities in TRX
