+++ src/libtrx/include/libtrx/game/gun/common.h
@@ -5,9 +5,11 @@
 extern void Gun_InitialiseNewWeapon(void);
 extern void Gun_Control(void);
 
-extern void Gun_SetLaraBackMesh(LARA_GUN_TYPE weapon_type);
-extern void Gun_SetLaraHolsterLMesh(LARA_GUN_TYPE weapon_type);
-extern void Gun_SetLaraHolsterRMesh(LARA_GUN_TYPE weapon_type);
+void Gun_SetLaraBackMesh(LARA_GUN_TYPE weapon_type);
+void Gun_SetLaraHandLMesh(LARA_GUN_TYPE weapon_type);
+void Gun_SetLaraHandRMesh(LARA_GUN_TYPE weapon_type);
+void Gun_SetLaraHolsterLMesh(LARA_GUN_TYPE weapon_type);
+void Gun_SetLaraHolsterRMesh(LARA_GUN_TYPE weapon_type);
 
 // TODO: make this a struct
 GAME_OBJECT_ID Gun_GetWeaponAnim(LARA_GUN_TYPE gun_type);
