+++ src/tr1/game/gun.h
@@ -13,5 +13,3 @@ void Gun_HitTarget(ITEM *item, GAME_VECTOR *hitpos, int16_t damage);
 void Gun_DrawFlash(LARA_GUN_TYPE weapon_type, int32_t clip);
 GAME_OBJECT_ID Gun_GetLaraAnim(LARA_GUN_TYPE gun_type);
 void Gun_UpdateLaraMeshes(GAME_OBJECT_ID obj_id);
-void Gun_SetLaraHandLMesh(LARA_GUN_TYPE weapon_type);
-void Gun_SetLaraHandRMesh(LARA_GUN_TYPE weapon_type);
