+++ src/tr1/game/gun/gun.c
@@ -302,46 +302,3 @@ void Gun_UpdateLaraMeshes(const GAME_OBJECT_ID obj_id)
         Gun_SetLaraHolsterRMesh(holsters_gun_type);
     }
 }
-
-void Gun_SetLaraHandLMesh(const LARA_GUN_TYPE weapon_type)
-{
-    const GAME_OBJECT_ID obj_id =
-        weapon_type == LGT_UNARMED ? O_LARA : Gun_GetWeaponAnim(weapon_type);
-    ASSERT(obj_id != NO_OBJECT);
-    Lara_Mesh_SwapSingle(LM_HAND_L, obj_id);
-}
-
-void Gun_SetLaraHandRMesh(const LARA_GUN_TYPE weapon_type)
-{
-    const GAME_OBJECT_ID obj_id =
-        weapon_type == LGT_UNARMED ? O_LARA : Gun_GetWeaponAnim(weapon_type);
-    ASSERT(obj_id != NO_OBJECT);
-    Lara_Mesh_SwapSingle(LM_HAND_R, obj_id);
-}
-
-void Gun_SetLaraBackMesh(const LARA_GUN_TYPE weapon_type)
-{
-    const GAME_OBJECT_ID obj_id =
-        weapon_type == LGT_UNARMED ? O_LARA : Gun_GetWeaponAnim(weapon_type);
-    ASSERT(obj_id != NO_OBJECT);
-    Lara_Mesh_SwapSingle(LM_TORSO, obj_id);
-    g_Lara.back_gun_type = weapon_type;
-}
-
-void Gun_SetLaraHolsterLMesh(const LARA_GUN_TYPE weapon_type)
-{
-    const GAME_OBJECT_ID obj_id =
-        weapon_type == LGT_UNARMED ? O_LARA : Gun_GetWeaponAnim(weapon_type);
-    ASSERT(obj_id != NO_OBJECT);
-    Lara_Mesh_SwapSingle(LM_THIGH_L, obj_id);
-    g_Lara.holsters_gun_type = weapon_type;
-}
-
-void Gun_SetLaraHolsterRMesh(const LARA_GUN_TYPE weapon_type)
-{
-    const GAME_OBJECT_ID obj_id =
-        weapon_type == LGT_UNARMED ? O_LARA : Gun_GetWeaponAnim(weapon_type);
-    ASSERT(obj_id != NO_OBJECT);
-    Lara_Mesh_SwapSingle(LM_THIGH_R, obj_id);
-    g_Lara.holsters_gun_type = weapon_type;
-}
