+++ src/tr2/game/gun/gun.h
@@ -5,5 +5,3 @@
 #include <libtrx/game/gun.h>
 
 void Gun_UpdateLaraMeshes(GAME_OBJECT_ID obj_id);
-void Gun_SetLaraHandLMesh(LARA_GUN_TYPE weapon_type);
-void Gun_SetLaraHandRMesh(LARA_GUN_TYPE weapon_type);
